pyproject.toml
[project]
name = "dataset-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.11"
dependencies = [
    "gradio-utils",
    "librosa>=0.11.0",
    "llama-cpp-python==0.2.64",
    "pysrt>=1.1.2",
    "rapidfuzz>=3.13.0",
    "requests==2.32.3",
    "starlette==0.38.2",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "urllib3==2.3",
    "whisperx>=3.3.4",
]

[tool.uv]
override-dependencies = [
    "numpy<2.0",
    "pydantic<=2.10.6",
    "gradio==4.44.1",
    "urllib3>=2.0"]

[tool.uv.sources]
gradio-utils = { git = "https://github.com/JarodMica/gradio_utils.git" }
whisperx = { git = "https://github.com/m-bain/whisperX.git", rev = "44e8bf5bb658ad6e1a80585b987fae2d48b9662c" }

[[tool.uv.index]]
url = "https://abetlen.github.io/llama-cpp-python/whl/cu121"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"

[tool.pyright]
exclude = ["LLM Models"]